requirements.txt
livekit-agents>=0.9.0
aiohttp>=3.8.0
livekit-plugins-openai>=0.10.15